Removed typecast

+++ happybase/util.py
@@ -49,7 +49,7 @@ class RepeatedTimer(object):
 
     def start(self):
         if not self.is_running:
-            self._timer = Timer(int(self.interval/1000), self._run)
+            self._timer = Timer(self.interval/1000, self._run)
             self._timer.start()
             self.is_running = True
